# src/openapi.yml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Auth API
  description: "Fursona"
components:
  securitySchemes:
    CookieAuth: # not supported by browser atm
      type: apiKey
      in: cookie
      name: token
servers:
  - url: /
  - url: /v1
paths:
...
  /login:
    post:
      summary: Log in a user
      description: Logs in a user with the provided credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  default: 'example@gmail.com'
                password:
                  type: string
                  default: '123'
      responses:
        '200':
          description: User logged in successfully
          headers:
            Set-Cookie:
              description: JWT token stored in a cookie
              schema:
                type: string
        '403':
          description: Unauthorized
        '404':
          description: User not found
...
  /admin:
      get:
        security:
          - CookieAuth: []
        summary: Check if user is an admin
        description: Determines whether the requesting user has admin privileges.
        responses:
          '200':
            description: User is an admin
          '403':
            description: User is not authorized
...
  /user:
    get:
      security:
        - CookieAuth: []
      summary: Get user records
      description: Retrieves user activity records.
      responses:
        '200':
          description: Successfully retrieved records
        '403':
          description: Unauthorized access
        '404':
          description: User not found
...